rename mocked scheduler to passthrough scheduler
